backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery==5.3.4
redis==5.0.1
pinecone-client==2.2.4
boto3==1.34.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
pdfminer.six==20221105
aiofiles==23.2.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
ibm-watson-machine-learning==1.0.362
ibm-granite-sdk==0.1.0